test: 파일 읽기 테스트 작성
