--- src/test/java/attendance/FileReader.java
@@ -1,4 +1,0 @@
-package attendance;
-
-public class FileReader {
-}
